:fire: Move all dataset processors to one file (`dataset_processors.py`)
